requirements.txt
blinker==1.7.0
click==8.1.7
easyocr==1.7.1
filelock==3.13.1
Flask==3.0.2
fsspec==2024.2.0
imageio==2.34.0
itsdangerous==2.1.2
Jinja2==3.1.3
lazy_loader==0.3
lxml==5.1.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.2.0
pyclipper==1.3.0.post5
pytesseract==0.3.10
python-bidi==0.4.2
python-docx==1.1.0
PyYAML==6.0.1
scikit-image==0.22.0
scipy==1.12.0
shapely==2.0.3
six==1.16.0
sympy==1.12
tifffile==2024.2.12
torch==2.2.1
torchvision==0.17.1
triton==2.2.0
typing_extensions==4.10.0
Werkzeug==3.0.1
